Add files via upload
This project focused on analyzing the performance of a YouTube channel using historical analytics data. The goal was to uncover trends in views, impressions, subscriber growth, and revenue while also identifying the best strategies for publishing content. To achieve this, I designed and developed an interactive dashboard in Tableau that enables stakeholders to explore key performance metrics and make data-driven publishing decisions.

Dashboard Features:

- KPI Overview (Top Cards)
- Trend Analysis (Line Chart)
- Video Lifecycle Analysis (Bar + Line Chart)
- Median Impressions segmented by inactive days (0–7, 8–30, 31–90, 181–365).
- Publishing Strategy (Heatmap)
- Best days of the week and months of the year to publish videos.
- Revenue Insights (Pie Chart & KPIs)

The final dashboard allows content creators and channel managers to:
1. Track overall channel performance.
2. Optimize publishing schedules based on data.
3. Understand the video lifecycle and focus on strategies that maximize impressions in the critical first 30 days.
4. Monitor revenue sources and identify opportunities to grow income streams.